+++ sop-java/src/main/kotlin/sop/Profile.kt
@@ -20,17 +20,22 @@ import sop.util.UTF8Util
  *   in the IETF namespace that begins with the string `draft-` should have semantics that hew as
  *   closely as possible to the referenced Internet Draft.
  * @param description a free-form description of the profile.
- * @see <a
- *   href="https://www.ietf.org/archive/id/draft-dkg-openpgp-stateless-cli-05.html#name-profile">
- *   SOP Spec - Profile</a>
+ * @param aliases list of optional profile alias names
+ * @see
+ *   [SOP Spec - Profile](https://www.ietf.org/archive/id/draft-dkg-openpgp-stateless-cli-05.html#name-profile)
  */
-data class Profile(val name: String, val description: Optional<String>) {
+data class Profile(
+    val name: String,
+    val description: Optional<String>,
+    val aliases: List<String> = listOf()
+) {
 
     @JvmOverloads
     constructor(
         name: String,
-        description: String? = null
-    ) : this(name, Optional.ofNullable(description?.trim()?.ifBlank { null }))
+        description: String? = null,
+        aliases: List<String> = listOf()
+    ) : this(name, Optional.ofNullable(description?.trim()?.ifBlank { null }), aliases)
 
     init {
         require(name.trim().isNotBlank()) { "Name cannot be empty." }
@@ -46,12 +51,32 @@ data class Profile(val name: String, val description: Optional<String>) {
     fun hasDescription() = description.isPresent
 
     /**
+     * Return a copy of this [Profile] with the aliases set to the given strings.
+     *
+     * @param alias one or more alias names
+     * @return profile with aliases
+     */
+    fun withAliases(vararg alias: String): Profile {
+        return Profile(name, description, alias.toList())
+    }
+
+    /**
      * Convert the profile into a String for displaying.
      *
      * @return string
      */
-    override fun toString(): String =
-        if (description.isEmpty) name else "$name: ${description.get()}"
+    override fun toString(): String = buildString {
+        append(name)
+        if (!description.isEmpty || !aliases.isEmpty()) {
+            append(":")
+        }
+        if (!description.isEmpty) {
+            append(" ${description.get()}")
+        }
+        if (!aliases.isEmpty()) {
+            append(" (aliases: ${aliases.joinToString(separator = ", ")})")
+        }
+    }
 
     companion object {
 
@@ -64,9 +89,21 @@ data class Profile(val name: String, val description: Optional<String>) {
         @JvmStatic
         fun parse(string: String): Profile {
             return if (string.contains(": ")) {
-                Profile(
-                    string.substring(0, string.indexOf(": ")),
-                    string.substring(string.indexOf(": ") + 2).trim())
+                val name = string.substring(0, string.indexOf(": "))
+                var description = string.substring(string.indexOf(": ") + 2).trim()
+                if (description.contains("(aliases: ")) {
+                    val aliases =
+                        description.substring(
+                            description.indexOf("(aliases: ") + 10, description.indexOf(")"))
+                    description = description.substring(0, description.indexOf("(aliases: ")).trim()
+                    Profile(name, description, aliases.split(", ").toList())
+                } else {
+                    if (description.isNotBlank()) {
+                        Profile(name, description)
+                    } else {
+                        Profile(name)
+                    }
+                }
             } else if (string.endsWith(":")) {
                 Profile(string.substring(0, string.length - 1))
             } else {